start working on dpo trainer + add s3 upload callback
