+++ requirements.txt
@@ -3,6 +3,7 @@ tensorboard
 datasets
 peft
 bitsandbytes
+trl
 webcolors
 pandas
 # flash-attn
